# packages/api-spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "The Music Repository API",
    "description": "The Music Repository API — the single source of truth for the HTTP contract.\nCompiles to OpenAPI 3.1 (`openapi.json`), which drives Orval (FE client + Zod) and the\nbackend Zod DTOs. Paths are declared ONLY here (never in NestJS decorators).",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/admin/redeem-codes": {
      "post": {
        "operationId": "createRedeemCode",
        "description": "Mint a gift/redeem code (staff only).",
        "parameters": [],
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedeemCodeCreated"
                }
              }
            }
          }
        },
        "tags": [
          "monetization"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRedeemCodeInput"
              }
            }
          }
        }
      }
    },
...
    "/me/redeem": {
      "post": {
        "operationId": "redeemCode",
        "description": "Redeem a gift code for the acting user → grants premium.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedeemResult"
                }
              }
            }
          }
        },
        "tags": [
          "monetization"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemCodeInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRedeemCodeInput": {
        "type": "object",
        "properties": {
          "tier": {
            "type": "string",
            "description": "Entitlement tier granted: \"premium\" (default) or \"pro\"."
          },
          "durationDays": {
            "type": "integer",
            "format": "int32",
            "description": "Days the granted premium lasts; omit for no expiry."
          },
          "uses": {
            "type": "integer",
            "format": "int32",
            "description": "How many times the code can be redeemed (default 1)."
          }
        },
        "description": "Body for minting a gift code (staff only)."
      },
...
      "RedeemCodeCreated": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "RedeemCodeInput": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "description": "Body for redeeming a code."
      },
      "RedeemResult": {
        "type": "object",
        "required": [
          "redeemed"
        ],
        "properties": {
          "redeemed": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string"
          }
        }
      },
...
    }
  }
}